nesting area computation standartization
