+++ RefactoringsExperiments/src/main/java/org/jetbrains/research/extractMethodExperiments/models/Fragment.java
@@ -21,6 +21,9 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.regex.Pattern;
 
+import static org.jetbrains.research.extractMethodExperiments.utils.feature.generators.DepthAnalyzer.getNestingArea;
+import static org.jetbrains.research.extractMethodExperiments.utils.feature.generators.DepthAnalyzer.getNestingDepth;
+
 public class Fragment {
     private MethodDeclaration methodDeclaration;
     /**
@@ -47,41 +50,6 @@ public class Fragment {
         this.logger = logger;
         this.methodArea = getNestingArea(methodDeclaration.toString());
         this.methodDepth = getNestingDepth(methodDeclaration.toString());
-    }
-
-
-    public static int[] getNestingDepths(String code) {
-        String[] lines = code.split("\n");
-        int depthInLine[] = new int[lines.length];
-        int currentDepth = 0; // current count
-
-        // Traverse the input strings
-        for (int i = 0; i < lines.length; i++) {
-            for (int j = 0; j < lines[i].length(); j++) {
-                if (lines[i].charAt(j) == '{') break;
-
-                if (lines[i].charAt(j) == '}') {
-                    currentDepth--;
-                }
-            }
-            depthInLine[i] = currentDepth;
-            for (int j = 0; j < lines[i].length(); j++) {
-                if (lines[i].charAt(j) == '{') currentDepth++;
-            }
-        }
-        return depthInLine;
-    }
-
-    public static int getNestingArea(String code) {
-        int area = 0;
-        for (int value : getNestingDepths(code)) area += value;
-        return area;
-    }
-
-    public static int getNestingDepth(String code) {
-        int depth = 0;
-        for (int value : getNestingDepths(code)) depth = Math.max(value, depth);
-        return depth;
     }
